ignore SIE_STATUS.NAK_REC in IRQ handler
The RP2040 can neither stop current transaction on NAK not raise an IRQ
on that event, so if device repeatedly NAKs transfers, the USB Host keeps
resending packet until it gets ACKed or ony other USB error occurs.
At the same time, if it was ACKed after a NAK, both NAK_REC and BUFF_STATUS
bits are set, in this case we should _not_ drop the data and clear the
stray NAK_REC.
RP2350 supports stop/IRQ on NAK so this event can be handled correctly there.